refactor: switch main functions to app.js to handle props properly / add gamelist page
